Each 'build teams' command now will uses a different seed.
@maikon
@jsuchy
@felipesere

I realized that I was using the same seed each time a 'build teams' request
was received and that meant the same order of random choices was being re-played.  This obviously doesn't look random if the member set remains the same.  It took me a while to work out how to do this.  I needed to work it through with Mateu as it's a fairly complicated topic.