Separate TeamAllocator modules between MaxSizeTeamAllocator and FixedTeamAllocator
This is an attempt to remove the duplication that I showed in my IPM in the RandomTeamAllocator
module.  I don't know if I have actually achieved my goal to remove all of the duplication
but I am postponing any more refactoring until I have fixed the MaxSizeTeam algorithm which
isn't currently evenly distributing.  This will have an effect on MaxSizeTeamAllocator method
assign_teams(...).
@maikon
@jsuchy